added an android service layer and piped everything up! and it builds (android unit tests dont work but we probably dont need them because of the jvm unit tests)

+++ wire4/src/commonTest/kotlin/mqtt/wire4/control/packet/PublishMessageTests.kt
@@ -31,7 +31,7 @@ class PublishMessageTests {
     @Test
     fun qos0AndPacketIdentifierThrowsIllegalArgumentException() {
         val fixed = FixedHeader(qos = QualityOfService.AT_MOST_ONCE)
-        val variable = VariableHeader(Name("t"), 2.toUShort())
+        val variable = VariableHeader(Name("t"), 2)
         try {
             PublishMessage(fixed, variable)
             fail()
